Fix warnings on Solaris 10/opteron reported by Yamamoto-san in [gfarm-developers:00470]
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@2847 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ gftool/gfarm_agent/gfarm_agent.c
@@ -1387,8 +1387,8 @@ display_env(int fd, int port)
 			break;
 		}			
 		fprintf(f, "GFARM_AGENT_PID=%d; export GFARM_AGENT_PID;\n",
-			pid);
-		fprintf(f, "echo Agent pid %d;\n", pid);
+			(int)pid);
+		fprintf(f, "echo Agent pid %d;\n", (int)pid);
 		break;
 	case C_SHELL_LIKE:
 		switch (server_type) {
@@ -1402,8 +1402,8 @@ display_env(int fd, int port)
 		default:
 			break;
 		}
-		fprintf(f, "setenv GFARM_AGENT_PID %d;\n", pid);
-		fprintf(f, "echo Agent pid %d;\n", pid);
+		fprintf(f, "setenv GFARM_AGENT_PID %d;\n", (int)pid);
+		fprintf(f, "echo Agent pid %d;\n", (int)pid);
 		break;
 	default:
 		break;
